create_message will return string now prospects_add_to_step added to Campaign move _send_email_work to Campaign
